esports_parser::coaches
Struct Coach

pub struct Coach {
    coach_name: String,
    top: String,
    jng: String,
    mid: String,
    bot: String,
    sup: String,
    team: String,
    weeklypoints: f64,
    totalpoints: f64,
}

Struct for Coach. This stores the Coach’s player names, coach name, lck team and both weekly/total points

Fields
coach_name: Stringtop: Stringjng: Stringmid: Stringbot: Stringsup: Stringteam: Stringweeklypoints: f64totalpoints: f64
Implementations

impl Coach

pub fn new( coach_name: String, top: String, jng: String, mid: String, bot: String, sup: String, team: String, ) -> Self

Constructor that builds a new empty Coach

pub fn get_coach_name(&self) -> &str

Getter for Coach. Returns coach_name

Examples

example_coach.teamname();
"fraudler1"

pub fn get_team(&self) -> &str

Getter for Coach. Returns team

Examples

example_coach.get_team();
"Dplus Kia"

pub fn get_totalpoints(&self) -> f64

Getter for Coach. Returns totalpoints

Examples

example_coach.get_totalpoints();
143.23

pub fn set_team(&mut self, new_team: String)

Setter for Coach. Sets a inputted team for Coach

Examples

example_coach.get_totalpoints("KT Rolster");
example_coach.team = KT Rolster

pub fn set_role(&mut self, role: &str, player_name: String)

Sets the pro player according to player’s role. It they are attempting to store to store a pro in the wrong role, it will output error

Examples

coach.set_role("top", siwoo);
Coach's top laner set to Siwoo.
OR
coach.set_role("jng", siwoo);
Invalid role jng

pub fn set_weekly(&mut self, pro_list: &[ProStats], team_list: &[Team])

Sets weekly score for Coach by finding all the players/teams in the coach’s team and totalling it.

pub fn set_total(&mut self)

Adds the current totalpoints with weeklypoints, updating totalpoints

pub fn load_from_csv(file_path: &str) -> Result<Vec<Coach>, Box<dyn Error>>

Loads coaches.csv to deseralize and store

pub fn remove_coach_by_name(coaches: &mut Vec<Coach>, name: &str) -> bool

Removes a inputted Coach from the coaches’ vector

Examples

Current Coaches: fraduler1, badname, notreal
remove_coach_by_name(coaches, fraduler1)
Current Coaches: badname, notreal

pub fn print_team_by_coach( coaches: &[Coach], players: &[ProStats], target_coach: &str, )

Prints inputted Coach’s team and scores

Examples

Name                     Kills    Deaths   Assists     CS    Weekly     Total
fraudler1                                                      0.00      0.00
Siwoo                       54        46       107   4113     44.33     44.33
Peanut                      58        66       212   4697     65.10     65.10
Faker                       51        37       103   4607     35.37     35.37
Jiwoo                      111        63       138   7441     63.67     63.67
Delight                     32        77       242    916     65.16     65.16
gen.g

pub fn print_all(coaches: &[Coach])

Prints all the coaches and their weekly/total points

Examples

All Coaches:
Coach Name:                Weekly Points        Total Points
fraudler1                           0.00                0.00
badname                             0.00                0.00

pub fn write_to_csv( coaches: &mut [Coach], file_path: &str, ) -> Result<(), Box<dyn Error>>

Store ALL of the Coaches’ data into coaches.csv. Resets weekly points to 0 to prepare for next week