esports_parser
Module stats

Structs

ProStats

A struct that stores the stats and fantasy points of each pro player

Team

Struct for each LCK Team, stores their wins and names

Functions

default_empty_string_to_f64 🔒

Turns an empty string to a f64 in order to avoid errors when deserializing

role_order 🔒

Matches a role with a usize to help with sorting by role

store_player

Parses through both pro_list.csv and data.csv and store the data into two vectors One vector is for ProStats for all the pro players and the other is for Teams for all the Teams in the LCK. All the stats are deserialized and stored into their structs. They are stored in alphabetical order and for ProStats, they are stored in order of role (top -> jng -> mid -> bot -> sup)

write_players_to_csv

Stores all the pro players to pro_list.csv. Resets everything to 0 except for totalpoints in order to reset for the next week.