esports_parser::stats
Struct ProStats

pub struct ProStats {
    position: String,
    playername: String,
    teamname: String,
    kills: f64,
    deaths: f64,
    assists: f64,
    total_cs: f64,
    league: String,
    weeklypoints: f64,
    totalpoints: f64,
    result: u32,
}

A struct that stores the stats and fantasy points of each pro player

Fields
position: Stringplayername: Stringteamname: Stringkills: f64deaths: f64assists: f64total_cs: f64league: Stringweeklypoints: f64totalpoints: f64result: u32
Implementations

impl ProStats

pub fn playername(&self) -> &str

Getter for ProStats. Returns playername

Examples

example_player.playername()
Siwoo

pub fn weeklypoints(&self) -> f64

Getter for ProStats. Returns weeklypoints

Examples

example_player.weeklypoints()
24.5

pub fn position(&self) -> &str

Getter for ProStats. Returns position

Examples

example_player.position()
top

pub fn update_totalpoints(&mut self)

Adds the current totalpoints with weeklypoints

Examples

example_player.update_totalpoints()
24.00 += 25.00
totalpoints = 29.00

pub fn calculate_points(&mut self)

Calculates pro playerโs points based off role and stats

Examples

example_player.calculate_points()
(example_player is top lane)
weeklypoints = (3 * 0.4) + (2 * 0.2) + (202 * 0.002) - (2 * 0.15)
weeklypoints = 1.7

pub fn print_all(players: &[ProStats])

Prints all of the pro players

Examples

Name                     Kills    Deaths   Assists     CS    Weekly     Total
BNK FEARX
Clear                       20        48        59   2885     18.36     18.36
Raptor                      32        51        67   2206     22.98     22.98
VicLa                       26        49        67   3261     15.54     15.54
Diable                      59        34        52   3710     30.16     30.16
Kellin                       8        54        96    335     22.01     22.01
etc...

pub fn print_by_name(players: &[ProStats], name: &str)

Searches through ProStats vector for the inputted playername

Examples

ProStats::print_by_name(players, Clear);
Name                     Kills    Deaths   Assists     CS    Weekly     Total
Clear                       20        48        59   2885     18.36     18.36

pub fn print_by_name_no_header(players: &[ProStats], name: &str)

Searches through ProStats vector for the inputted playername, but DOES NOT print header. Used more as a helper function for Coaches.

Examples

ProStats::print_by_name_no_header(players, Clear);
Clear                       20        48        59   2885     18.36     18.36

pub fn print_by_team(players: &[ProStats], teamname: &str)

Searches through ProStats vector for the inputted teamname

Examples

ProStats::print_by_team(players, BNK FearX);
Name                     Kills    Deaths   Assists     CS    Weekly     Total
BNK FEARX
Clear                       20        48        59   2885     18.36     18.36
Raptor                      32        51        67   2206     22.98     22.98
VicLa                       26        49        67   3261     15.54     15.54
Diable                      59        34        52   3710     30.16     30.16
Kellin                       8        54        96    335     22.01     22.01

pub fn print_top_role(players: &[ProStats])

Prints the top weekly scorers per role

Examples

ProStats::print_top_role(players);
Top Players by Role:
Name                     Kills    Deaths   Assists     CS    Weekly     Total
Morgan                   10043        40        71   3860   4033.11   4033.11
GIDEON                      69        55       199   5481     70.12     70.12
Zeka                        93        51       148   6446     58.87     58.87
Viper                      134        48       146   6940     76.32     76.32
Delight                     32        77       242    916     65.16     65.16

pub fn print_header()

Helper function that prints the header for each of the pro stats printer functions

Examples

print_header()
Name                     Kills    Deaths   Assists     CS    Weekly     Total

pub fn print_player(player: &ProStats)

Helper function that prints the pro stats for a player

Examples

GIDEON                      69        55       199   5481     70.12     70.12